+++ src/main/java/racingcar/reader/RacingInfo.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.reader;
 
 import java.util.List;
