--- src/test/java/racingcar/RacingGameManagerTest.java
@@ -1,21 +1,0 @@
-package racingcar;
-
-import static org.mockito.Mockito.mockStatic;
-import camp.nextstep.edu.missionutils.Console;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.mockito.MockedStatic;
-
-public class RacingGameManagerTest {
-	MockedStatic<Console> consoleMockedStatic;
-
-	@BeforeEach
-	public void setUp() throws Exception {
-		consoleMockedStatic = mockStatic(Console.class);
-	}
-
-	@AfterEach
-	public void tearDown() throws Exception {
-		consoleMockedStatic.close();
-	}
-}
