+++ src/test/java/racingcar/reader/CarNameReaderTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.reader;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.MockedStatic;
+import racingcar.ConsoleDisplay;
 import racingcar.reader.CarNameReader;
 
 public class CarNameReaderTest {
