+++ src/test/java/racingcar/reader/RacingInfoReaderTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.reader;
 
 import camp.nextstep.edu.missionutils.Console;
 import org.junit.jupiter.api.AfterEach;
@@ -8,7 +8,10 @@ import org.junit.jupiter.api.Test;
 import org.mockito.InOrder;
 import org.mockito.MockedStatic;
 import org.mockito.Mockito;
+import racingcar.ConsoleDisplay;
 import racingcar.reader.CarNameReader;
+import racingcar.reader.RacingInfo;
+import racingcar.reader.RacingInfoReader;
 import racingcar.reader.TurnValueReader;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.*;
